Refactor monolithic files: application.cpp, about_window.cpp, palette_common.cpp
- Extracted `MainFrame` from `source/app/application.cpp` to `source/ui/main_frame.{h,cpp}`.
- Refactored `AboutWindow` by moving `TetrisPanel`, `SnakePanel`, and `GamePanel` to `source/ui/games/`.
- Decomposed `source/palette/palette_common.cpp` into `PalettePanel`, `BrushSizePanel`, `BrushToolPanel`, `BrushThicknessPanel`, and `BrushButton` in `source/palette/`.
- Updated `CMakeLists.txt` to reflect file structure changes.
- Preserved `palette_common.h` as an aggregate header for backward compatibility.
