+++ source/app/application.cpp
@@ -45,54 +45,13 @@
 #include <wx/snglinst.h>
 #include <spdlog/spdlog.h>
 
+#include "ui/main_frame.h"
+
 #if defined(__LINUX__) || defined(__WINDOWS__)
 	#include <GL/glut.h>
 #endif
 
 #include "../brushes/icon/editor_icon.xpm"
-
-BEGIN_EVENT_TABLE(MainFrame, wxFrame)
-EVT_CLOSE(MainFrame::OnExit)
-
-// Update check complete
-#ifdef _USE_UPDATER_
-EVT_ON_UPDATE_CHECK_FINISHED(wxID_ANY, MainFrame::OnUpdateReceived)
-#endif
-EVT_ON_UPDATE_MENUS(wxID_ANY, MainFrame::OnUpdateMenus)
-
-// Idle event handler
-EVT_IDLE(MainFrame::OnIdle)
-END_EVENT_TABLE()
-
-BEGIN_EVENT_TABLE(MapWindow, wxPanel)
-EVT_SIZE(MapWindow::OnSize)
-
-EVT_COMMAND_SCROLL_TOP(MAP_WINDOW_HSCROLL, MapWindow::OnScroll)
-EVT_COMMAND_SCROLL_BOTTOM(MAP_WINDOW_HSCROLL, MapWindow::OnScroll)
-EVT_COMMAND_SCROLL_THUMBTRACK(MAP_WINDOW_HSCROLL, MapWindow::OnScroll)
-EVT_COMMAND_SCROLL_LINEUP(MAP_WINDOW_HSCROLL, MapWindow::OnScrollLineUp)
-EVT_COMMAND_SCROLL_LINEDOWN(MAP_WINDOW_HSCROLL, MapWindow::OnScrollLineDown)
-EVT_COMMAND_SCROLL_PAGEUP(MAP_WINDOW_HSCROLL, MapWindow::OnScrollPageUp)
-EVT_COMMAND_SCROLL_PAGEDOWN(MAP_WINDOW_HSCROLL, MapWindow::OnScrollPageDown)
-
-EVT_COMMAND_SCROLL_TOP(MAP_WINDOW_VSCROLL, MapWindow::OnScroll)
-EVT_COMMAND_SCROLL_BOTTOM(MAP_WINDOW_VSCROLL, MapWindow::OnScroll)
-EVT_COMMAND_SCROLL_THUMBTRACK(MAP_WINDOW_VSCROLL, MapWindow::OnScroll)
-EVT_COMMAND_SCROLL_LINEUP(MAP_WINDOW_VSCROLL, MapWindow::OnScrollLineUp)
-EVT_COMMAND_SCROLL_LINEDOWN(MAP_WINDOW_VSCROLL, MapWindow::OnScrollLineDown)
-EVT_COMMAND_SCROLL_PAGEUP(MAP_WINDOW_VSCROLL, MapWindow::OnScrollPageUp)
-EVT_COMMAND_SCROLL_PAGEDOWN(MAP_WINDOW_VSCROLL, MapWindow::OnScrollPageDown)
-
-EVT_BUTTON(MAP_WINDOW_GEM, MapWindow::OnGem)
-END_EVENT_TABLE()
-
-BEGIN_EVENT_TABLE(MapScrollBar, wxScrollBar)
-EVT_KEY_DOWN(MapScrollBar::OnKey)
-EVT_KEY_UP(MapScrollBar::OnKey)
-EVT_CHAR(MapScrollBar::OnKey)
-EVT_SET_FOCUS(MapScrollBar::OnFocus)
-EVT_MOUSEWHEEL(MapScrollBar::OnWheel)
-END_EVENT_TABLE()
 
 wxIMPLEMENT_APP(Application);
 
@@ -376,317 +335,4 @@ bool Application::ParseCommandLineMap(wxString& fileName) {
 		}
 	}
 	return false;
-}
-
-MainFrame::MainFrame(const wxString& title, const wxPoint& pos, const wxSize& size) :
-	wxFrame((wxFrame*)nullptr, -1, title, pos, size, wxDEFAULT_FRAME_STYLE) {
-	// Receive idle events
-	SetExtraStyle(wxWS_EX_PROCESS_IDLE);
-
-#if wxCHECK_VERSION(3, 1, 0) // 3.1.0 or higher
-	// Make sure ShowFullScreen() uses the full screen API on macOS
-	EnableFullScreenView(true);
-#endif
-
-	// Creates the file-dropdown menu
-	menu_bar = newd MainMenuBar(this);
-	wxArrayString warnings;
-	wxString error;
-
-	wxFileName filename;
-	filename.Assign(FileSystem::GetFoundDataDirectory() + "menubar.xml");
-	if (!filename.FileExists()) {
-		filename = FileName(FileSystem::GetDataDirectory() + "menubar.xml");
-	}
-
-	if (!menu_bar->Load(filename, warnings, error)) {
-		wxLogError(wxString() + "Could not load menubar.xml, editor will NOT be able to show its menu.\n");
-	}
-
-	wxStatusBar* statusbar = CreateStatusBar();
-	statusbar->SetFieldsCount(4);
-	SetStatusText(wxString("Welcome to ") << __W_RME_APPLICATION_NAME__ << " " << __W_RME_VERSION__);
-
-	// Le sizer
-	g_gui.aui_manager = newd wxAuiManager(this);
-	g_gui.tabbook = newd MapTabbook(this, wxID_ANY);
-
-	tool_bar = newd MainToolBar(this, g_gui.aui_manager);
-
-	g_gui.aui_manager->AddPane(g_gui.tabbook, wxAuiPaneInfo().CenterPane().Floatable(false).CloseButton(false).PaneBorder(false));
-
-	g_gui.aui_manager->Update();
-
-	UpdateMenubar();
-}
-
-MainFrame::~MainFrame() = default;
-
-void MainFrame::OnIdle(wxIdleEvent& event) {
-	if (menu_bar) {
-		menu_bar->Update();
-	}
-	if (tool_bar) {
-		tool_bar->UpdateButtons();
-	}
-	event.RequestMore();
-	event.Skip();
-}
-
-#ifdef _USE_UPDATER_
-void MainFrame::OnUpdateReceived(wxCommandEvent& event) {
-	std::string data = *(std::string*)event.GetClientData();
-	delete (std::string*)event.GetClientData();
-	size_t first_colon = data.find(':');
-	size_t second_colon = data.find(':', first_colon + 1);
-
-	if (first_colon == std::string::npos || second_colon == std::string::npos) {
-		return;
-	}
-
-	std::string update = data.substr(0, first_colon);
-	std::string verstr = data.substr(first_colon + 1, second_colon - first_colon - 1);
-	std::string url = (second_colon == data.size() ? "" : data.substr(second_colon + 1));
-
-	if (update == "yes") {
-		int ret = DialogUtil::PopupDialog(
-			"Update Notice",
-			wxString("There is a newd update available (") << wxstr(verstr) << "). Do you want to go to the website and download it?",
-			wxYES | wxNO,
-			"I don't want any update notices",
-			Config::AUTOCHECK_FOR_UPDATES
-		);
-		if (ret == wxID_YES) {
-			::wxLaunchDefaultBrowser(wxstr(url), wxBROWSER_NEW_WINDOW);
-		}
-	}
-}
-#endif
-
-void MainFrame::OnUpdateMenus(wxCommandEvent&) {
-	UpdateMenubar();
-	g_gui.UpdateMinimap(true);
-	g_gui.UpdateTitle();
-}
-
-#ifdef __WINDOWS__
-bool MainFrame::MSWTranslateMessage(WXMSG* msg) {
-	if (g_hotkeys.AreHotkeysEnabled()) {
-		if (wxFrame::MSWTranslateMessage(msg)) {
-			return true;
-		}
-	} else {
-		if (wxWindow::MSWTranslateMessage(msg)) {
-			return true;
-		}
-	}
-	return false;
-}
-#endif
-
-void MainFrame::UpdateMenubar() {
-	menu_bar->Update();
-	tool_bar->UpdateButtons();
-}
-
-bool MainFrame::DoQueryClose() {
-	Editor* editor = g_gui.GetCurrentEditor();
-	if (editor) {
-		if (editor->live_manager.IsLive()) {
-			long ret = DialogUtil::PopupDialog(
-				"Must Close Server",
-				wxString("You are currently connected to a live server, to close this map the connection must be severed."),
-				wxOK | wxCANCEL
-			);
-			if (ret == wxID_OK) {
-				editor->live_manager.CloseServer();
-			} else {
-				return false;
-			}
-		}
-	}
-	return true;
-}
-
-bool MainFrame::DoQuerySaveTileset(bool doclose) {
-
-	if (!g_materials.needSave()) {
-		// skip dialog when there is nothing to save
-		return true;
-	}
-
-	long ret = DialogUtil::PopupDialog(
-		"Export tileset",
-		"Do you want to export your tileset changes before exiting?",
-		wxYES | wxNO | wxCANCEL
-	);
-
-	if (ret == wxID_NO) {
-		// "no" - exit without saving
-		return true;
-	} else if (ret == wxID_CANCEL) {
-		// "cancel" - just close the dialog
-		return false;
-	}
-
-	// "yes" button was pressed, open tileset exporting dialog
-	if (g_gui.GetCurrentEditor()) {
-		ExportTilesetsWindow dlg(this, *g_gui.GetCurrentEditor());
-		dlg.ShowModal();
-		dlg.Destroy();
-	}
-
-	return !g_materials.needSave();
-}
-
-bool MainFrame::DoQuerySave(bool doclose) {
-	if (!g_gui.IsEditorOpen()) {
-		return true;
-	}
-
-	if (!DoQuerySaveTileset()) {
-		return false;
-	}
-
-	Editor& editor = *g_gui.GetCurrentEditor();
-	if (editor.live_manager.IsClient()) {
-		long ret = DialogUtil::PopupDialog(
-			"Disconnect",
-			"Do you want to disconnect?",
-			wxYES | wxNO
-		);
-
-		if (ret != wxID_YES) {
-			return false;
-		}
-
-		editor.live_manager.CloseServer();
-		return DoQuerySave(doclose);
-	} else if (editor.live_manager.IsServer()) {
-		long ret = DialogUtil::PopupDialog(
-			"Shutdown",
-			"Do you want to shut down the server? (any clients will be disconnected)",
-			wxYES | wxNO
-		);
-
-		if (ret != wxID_YES) {
-			return false;
-		}
-
-		editor.live_manager.CloseServer();
-		return DoQuerySave(doclose);
-	} else if (g_gui.ShouldSave()) {
-		long ret = DialogUtil::PopupDialog(
-			"Save changes",
-			"Do you want to save your changes to \"" + wxstr(g_gui.GetCurrentMap().getName()) + "\"?",
-			wxYES | wxNO | wxCANCEL
-		);
-
-		if (ret == wxID_YES) {
-			if (g_gui.GetCurrentMap().hasFile()) {
-				g_gui.SaveCurrentMap(true);
-			} else {
-				wxFileDialog file(this, "Save...", "", "", "*.otbm", wxFD_SAVE | wxFD_OVERWRITE_PROMPT);
-				int32_t result = file.ShowModal();
-				if (result == wxID_OK) {
-					g_gui.SaveCurrentMap(file.GetPath(), true);
-				} else {
-					return false;
-				}
-			}
-		} else if (ret == wxID_CANCEL) {
-			return false;
-		}
-	}
-
-	if (doclose) {
-		UnnamedRenderingLock();
-		g_gui.CloseCurrentEditor();
-	}
-
-	return true;
-}
-
-bool MainFrame::DoQueryImportCreatures() {
-	if (g_creatures.hasMissing()) {
-		long ret = DialogUtil::PopupDialog("Missing creatures", "There are missing creatures and/or NPC in the editor, do you want to load them from an OT monster/npc file?", wxYES | wxNO);
-		if (ret == wxID_YES) {
-			do {
-				wxFileDialog dlg(g_gui.root, "Import monster/npc file", "", "", "*.xml", wxFD_OPEN | wxFD_MULTIPLE | wxFD_FILE_MUST_EXIST);
-				if (dlg.ShowModal() == wxID_OK) {
-					wxArrayString paths;
-					dlg.GetPaths(paths);
-					for (uint32_t i = 0; i < paths.GetCount(); ++i) {
-						wxString error;
-						wxArrayString warnings;
-						bool ok = g_creatures.importXMLFromOT(FileName(paths[i]), error, warnings);
-						if (ok) {
-							DialogUtil::ListDialog("Monster loader errors", warnings);
-						} else {
-							wxMessageBox("Error OT data file \"" + paths[i] + "\".\n" + error, "Error", wxOK | wxICON_INFORMATION, g_gui.root);
-						}
-					}
-				} else {
-					break;
-				}
-			} while (g_creatures.hasMissing());
-		}
-	}
-	g_gui.RefreshPalettes();
-	return true;
-}
-
-void MainFrame::UpdateFloorMenu() {
-	menu_bar->UpdateFloorMenu();
-}
-
-bool MainFrame::LoadMap(FileName name) {
-	return g_gui.LoadMap(name);
-}
-
-void MainFrame::OnExit(wxCloseEvent& event) {
-	// clicking 'x' button
-
-	// do you want to save map changes?
-	while (g_gui.IsEditorOpen()) {
-		if (!DoQuerySave()) {
-			if (event.CanVeto()) {
-				event.Veto();
-				return;
-			} else {
-				break;
-			}
-		}
-	}
-	g_palettes.DestroyPalettes();
-	g_minimap.Destroy();
-	g_search.HideSearchWindow();
-	g_layout.SavePerspective();
-
-	g_gui.aui_manager->UnInit();
-	((Application&)wxGetApp()).Unload();
-	Destroy();
-}
-
-void MainFrame::AddRecentFile(const FileName& file) {
-	menu_bar->recentFilesManager.AddFile(file);
-}
-
-void MainFrame::LoadRecentFiles() {
-	menu_bar->recentFilesManager.Load(&g_settings.getConfigObject());
-}
-
-void MainFrame::SaveRecentFiles() {
-	menu_bar->recentFilesManager.Save(&g_settings.getConfigObject());
-}
-
-std::vector<wxString> MainFrame::GetRecentFiles() {
-	return menu_bar->recentFilesManager.GetFiles();
-}
-
-void MainFrame::PrepareDC(wxDC& dc) {
-	dc.SetLogicalOrigin(0, 0);
-	dc.SetAxisOrientation(1, 0);
-	dc.SetUserScale(1.0, 1.0);
-	dc.SetMapMode(wxMM_TEXT);
 }
