+++ source/palette/brush_tool_panel.cpp
@@ -16,372 +16,12 @@
 //////////////////////////////////////////////////////////////////////
 
 #include "app/main.h"
-
-#include "palette/palette_common.h"
-#include "brushes/brush.h"
+#include "palette/brush_tool_panel.h"
+#include "palette/managers/palette_manager.h"
 #include "brushes/managers/brush_manager.h"
-#include "game/sprites.h"
-#include "ui/gui.h"
-#include "ui/controls/item_buttons.h"
-#include "app/application.h"
-#include "palette/palette_waypoints.h"
-#include "palette/managers/palette_manager.h" // Added for g_palettes
-
-// ============================================================================
-// Palette Panel
-
-BEGIN_EVENT_TABLE(PalettePanel, wxPanel)
-EVT_TIMER(PALETTE_DELAYED_REFRESH_TIMER, WaypointPalettePanel::OnRefreshTimer)
-END_EVENT_TABLE()
-
-PalettePanel::PalettePanel(wxWindow* parent, wxWindowID id, long style) :
-	wxPanel(parent, id, wxDefaultPosition, wxDefaultSize, style),
-	refresh_timer(this, PALETTE_DELAYED_REFRESH_TIMER),
-	last_brush_size(0) {
-	////
-}
-
-PalettePanel::~PalettePanel() {
-	////
-}
-
-PaletteWindow* PalettePanel::GetParentPalette() const {
-	const wxWindow* w = this;
-	while ((w = w->GetParent()) && dynamic_cast<const PaletteWindow*>(w) == nullptr)
-		;
-	return const_cast<PaletteWindow*>(static_cast<const PaletteWindow*>(w));
-}
-
-void PalettePanel::InvalidateContents() {
-	for (ToolBarList::iterator iter = tool_bars.begin(); iter != tool_bars.end(); ++iter) {
-		(*iter)->InvalidateContents();
-	}
-}
-
-void PalettePanel::LoadCurrentContents() {
-	for (ToolBarList::iterator iter = tool_bars.begin(); iter != tool_bars.end(); ++iter) {
-		(*iter)->OnSwitchIn();
-	}
-	Fit();
-}
-
-void PalettePanel::LoadAllContents() {
-	for (ToolBarList::iterator iter = tool_bars.begin(); iter != tool_bars.end(); ++iter) {
-		(*iter)->LoadAllContents();
-	}
-}
-
-void PalettePanel::AddToolPanel(PalettePanel* panel) {
-	wxSizer* sp_sizer = newd wxStaticBoxSizer(wxVERTICAL, this, panel->GetName());
-	sp_sizer->Add(panel, 0, wxEXPAND);
-	GetSizer()->Add(sp_sizer, 0, wxEXPAND);
-
-	// GetSizer()->SetDimension(wxDefaultCoord, wxDefaultCoord, wxDefaultCoord, wxDefaultCoord);
-	// GetSizer()->Layout();
-	Fit();
-	//
-
-	tool_bars.push_back(panel);
-}
-
-void PalettePanel::SetToolbarIconSize(bool large_icons) {
-	for (ToolBarList::iterator iter = tool_bars.begin(); iter != tool_bars.end(); ++iter) {
-		(*iter)->SetToolbarIconSize(large_icons);
-	}
-}
-
-wxString PalettePanel::GetName() const {
-	switch (GetType()) {
-		case TILESET_TERRAIN:
-			return "Terrain Palette";
-		case TILESET_DOODAD:
-			return "Doodad Palette";
-		case TILESET_ITEM:
-			return "Item Palette";
-		case TILESET_COLLECTION:
-			return "Collections Palette";
-		case TILESET_CREATURE:
-			return "Creature Palette";
-		case TILESET_HOUSE:
-			return "House Palette";
-		case TILESET_RAW:
-			return "RAW Palette";
-		case TILESET_WAYPOINT:
-			return "Waypoint Palette";
-		case TILESET_UNKNOWN:
-			return "Unknown";
-	}
-	return wxEmptyString;
-}
-
-PaletteType PalettePanel::GetType() const {
-	return TILESET_UNKNOWN;
-}
-
-Brush* PalettePanel::GetSelectedBrush() const {
-	return nullptr;
-}
-
-int PalettePanel::GetSelectedBrushSize() const {
-	return 0;
-}
-
-void PalettePanel::SelectFirstBrush() {
-	// Do nothing
-}
-
-bool PalettePanel::SelectBrush(const Brush* whatbrush) {
-	return false;
-}
-
-void PalettePanel::OnUpdateBrushSize(BrushShape shape, int size) {
-	for (ToolBarList::iterator iter = tool_bars.begin(); iter != tool_bars.end(); ++iter) {
-		(*iter)->OnUpdateBrushSize(shape, size);
-	}
-}
-
-void PalettePanel::OnSwitchIn() {
-	for (ToolBarList::iterator iter = tool_bars.begin(); iter != tool_bars.end(); ++iter) {
-		(*iter)->OnSwitchIn();
-	}
-	g_palettes.ActivatePalette(GetParentPalette());
-	g_brush_manager.SetBrushSize(last_brush_size);
-}
-
-void PalettePanel::OnSwitchOut() {
-	last_brush_size = g_brush_manager.GetBrushSize();
-	for (ToolBarList::iterator iter = tool_bars.begin(); iter != tool_bars.end(); ++iter) {
-		(*iter)->OnSwitchOut();
-	}
-}
-
-void PalettePanel::OnUpdate() {
-	for (ToolBarList::iterator iter = tool_bars.begin(); iter != tool_bars.end(); ++iter) {
-		(*iter)->OnUpdate();
-	}
-}
-
-void PalettePanel::RefreshOtherPalettes() {
-	refresh_timer.Start(100, true);
-}
-
-void PalettePanel::OnRefreshTimer(wxTimerEvent&) {
-	g_palettes.RefreshOtherPalettes(GetParentPalette());
-}
-
-// ============================================================================
-// Size Page
-
-BEGIN_EVENT_TABLE(BrushSizePanel, wxPanel)
-EVT_TOGGLEBUTTON(PALETTE_BRUSHSHAPE_SQUARE, BrushSizePanel::OnClickSquareBrush)
-EVT_TOGGLEBUTTON(PALETTE_BRUSHSHAPE_CIRCLE, BrushSizePanel::OnClickCircleBrush)
-
-EVT_TOGGLEBUTTON(PALETTE_TERRAIN_BRUSHSIZE_0, BrushSizePanel::OnClickBrushSize0)
-EVT_TOGGLEBUTTON(PALETTE_TERRAIN_BRUSHSIZE_1, BrushSizePanel::OnClickBrushSize1)
-EVT_TOGGLEBUTTON(PALETTE_TERRAIN_BRUSHSIZE_2, BrushSizePanel::OnClickBrushSize2)
-EVT_TOGGLEBUTTON(PALETTE_TERRAIN_BRUSHSIZE_4, BrushSizePanel::OnClickBrushSize4)
-EVT_TOGGLEBUTTON(PALETTE_TERRAIN_BRUSHSIZE_6, BrushSizePanel::OnClickBrushSize6)
-EVT_TOGGLEBUTTON(PALETTE_TERRAIN_BRUSHSIZE_8, BrushSizePanel::OnClickBrushSize8)
-EVT_TOGGLEBUTTON(PALETTE_TERRAIN_BRUSHSIZE_11, BrushSizePanel::OnClickBrushSize11)
-END_EVENT_TABLE()
-
-BrushSizePanel::BrushSizePanel(wxWindow* parent) :
-	PalettePanel(parent, wxID_ANY),
-	loaded(false),
-	large_icons(true),
-	brushshapeSquareButton(nullptr),
-	brushshapeCircleButton(nullptr),
-	brushsize0Button(nullptr),
-	brushsize1Button(nullptr),
-	brushsize2Button(nullptr),
-	brushsize4Button(nullptr),
-	brushsize6Button(nullptr),
-	brushsize8Button(nullptr),
-	brushsize11Button(nullptr) {
-	////
-}
-
-void BrushSizePanel::InvalidateContents() {
-	if (loaded) {
-		DestroyChildren();
-		SetSizer(nullptr);
-
-		brushshapeSquareButton = brushshapeCircleButton = brushsize0Button = brushsize1Button = brushsize2Button = brushsize4Button = brushsize6Button = brushsize8Button = brushsize11Button = nullptr;
-
-		loaded = false;
-	}
-}
-
-void BrushSizePanel::LoadCurrentContents() {
-	LoadAllContents();
-}
-
-void BrushSizePanel::LoadAllContents() {
-	if (loaded) {
-		return;
-	}
-
-	wxSizer* size_sizer = newd wxBoxSizer(wxVERTICAL);
-	;
-	wxSizer* sub_sizer = newd wxBoxSizer(wxHORIZONTAL);
-	RenderSize render_size;
-
-	if (large_icons) {
-		// 32x32
-		render_size = RENDER_SIZE_32x32;
-	} else {
-		// 16x16
-		render_size = RENDER_SIZE_16x16;
-	}
-
-	sub_sizer->Add(brushshapeSquareButton = newd DCButton(this, PALETTE_BRUSHSHAPE_SQUARE, wxDefaultPosition, DC_BTN_TOGGLE, render_size, EDITOR_SPRITE_BRUSH_SD_9x9));
-	brushshapeSquareButton->SetToolTip("Square brush");
-
-	sub_sizer->Add(brushshapeCircleButton = newd DCButton(this, PALETTE_BRUSHSHAPE_CIRCLE, wxDefaultPosition, DC_BTN_TOGGLE, render_size, EDITOR_SPRITE_BRUSH_CD_9x9));
-	brushshapeCircleButton->SetToolTip("Circle brush");
-	brushshapeSquareButton->SetValue(true);
-
-	if (large_icons) {
-		sub_sizer->AddSpacer(36);
-	} else {
-		sub_sizer->AddSpacer(18);
-	}
-
-	sub_sizer->Add(brushsize0Button = newd DCButton(this, PALETTE_TERRAIN_BRUSHSIZE_0, wxDefaultPosition, DC_BTN_TOGGLE, render_size, EDITOR_SPRITE_BRUSH_CD_1x1));
-	brushsize0Button->SetToolTip("Brush size 1");
-	brushsize0Button->SetValue(true);
-
-	sub_sizer->Add(brushsize1Button = newd DCButton(this, PALETTE_TERRAIN_BRUSHSIZE_1, wxDefaultPosition, DC_BTN_TOGGLE, render_size, EDITOR_SPRITE_BRUSH_SD_3x3));
-	brushsize1Button->SetToolTip("Brush size 2");
-
-	if (large_icons) {
-		size_sizer->Add(sub_sizer);
-		sub_sizer = newd wxBoxSizer(wxHORIZONTAL);
-	}
-
-	sub_sizer->Add(brushsize2Button = newd DCButton(this, PALETTE_TERRAIN_BRUSHSIZE_2, wxDefaultPosition, DC_BTN_TOGGLE, render_size, EDITOR_SPRITE_BRUSH_SD_5x5));
-	brushsize2Button->SetToolTip("Brush size 3");
-
-	sub_sizer->Add(brushsize4Button = newd DCButton(this, PALETTE_TERRAIN_BRUSHSIZE_4, wxDefaultPosition, DC_BTN_TOGGLE, render_size, EDITOR_SPRITE_BRUSH_SD_7x7));
-	brushsize4Button->SetToolTip("Brush size 5");
-
-	sub_sizer->Add(brushsize6Button = newd DCButton(this, PALETTE_TERRAIN_BRUSHSIZE_6, wxDefaultPosition, DC_BTN_TOGGLE, render_size, EDITOR_SPRITE_BRUSH_SD_9x9));
-	brushsize6Button->SetToolTip("Brush size 7");
-
-	sub_sizer->Add(brushsize8Button = newd DCButton(this, PALETTE_TERRAIN_BRUSHSIZE_8, wxDefaultPosition, DC_BTN_TOGGLE, render_size, EDITOR_SPRITE_BRUSH_SD_15x15));
-	brushsize8Button->SetToolTip("Brush size 9");
-
-	sub_sizer->Add(brushsize11Button = newd DCButton(this, PALETTE_TERRAIN_BRUSHSIZE_11, wxDefaultPosition, DC_BTN_TOGGLE, render_size, EDITOR_SPRITE_BRUSH_SD_19x19));
-	brushsize11Button->SetToolTip("Brush size 12");
-
-	size_sizer->Add(sub_sizer);
-	SetSizerAndFit(size_sizer);
-
-	loaded = true;
-}
-
-wxString BrushSizePanel::GetName() const {
-	return "Brush Size";
-}
-
-void BrushSizePanel::SetToolbarIconSize(bool d) {
-	InvalidateContents();
-	large_icons = d;
-}
-
-void BrushSizePanel::OnSwitchIn() {
-	LoadCurrentContents();
-}
-
-void BrushSizePanel::OnUpdateBrushSize(BrushShape shape, int size) {
-	if (shape == BRUSHSHAPE_SQUARE) {
-		brushshapeCircleButton->SetValue(false);
-		brushshapeSquareButton->SetValue(true);
-
-		brushsize0Button->SetSprite(EDITOR_SPRITE_BRUSH_CD_1x1);
-		brushsize1Button->SetSprite(EDITOR_SPRITE_BRUSH_SD_3x3);
-		brushsize2Button->SetSprite(EDITOR_SPRITE_BRUSH_SD_5x5);
-		brushsize4Button->SetSprite(EDITOR_SPRITE_BRUSH_SD_7x7);
-		brushsize6Button->SetSprite(EDITOR_SPRITE_BRUSH_SD_9x9);
-		brushsize8Button->SetSprite(EDITOR_SPRITE_BRUSH_SD_15x15);
-		brushsize11Button->SetSprite(EDITOR_SPRITE_BRUSH_SD_19x19);
-	} else {
-		brushshapeSquareButton->SetValue(false);
-		brushshapeCircleButton->SetValue(true);
-
-		brushsize0Button->SetSprite(EDITOR_SPRITE_BRUSH_CD_1x1);
-		brushsize1Button->SetSprite(EDITOR_SPRITE_BRUSH_CD_3x3);
-		brushsize2Button->SetSprite(EDITOR_SPRITE_BRUSH_CD_5x5);
-		brushsize4Button->SetSprite(EDITOR_SPRITE_BRUSH_CD_7x7);
-		brushsize6Button->SetSprite(EDITOR_SPRITE_BRUSH_CD_9x9);
-		brushsize8Button->SetSprite(EDITOR_SPRITE_BRUSH_CD_15x15);
-		brushsize11Button->SetSprite(EDITOR_SPRITE_BRUSH_CD_19x19);
-	}
-
-	if (brushsize0Button) {
-		brushsize0Button->SetValue(false);
-	}
-	if (brushsize1Button) {
-		brushsize1Button->SetValue(false);
-	}
-	if (brushsize2Button) {
-		brushsize2Button->SetValue(false);
-	}
-	if (brushsize4Button) {
-		brushsize4Button->SetValue(false);
-	}
-	if (brushsize6Button) {
-		brushsize6Button->SetValue(false);
-	}
-	if (brushsize8Button) {
-		brushsize8Button->SetValue(false);
-	}
-	if (brushsize11Button) {
-		brushsize11Button->SetValue(false);
-	}
-
-	switch (size) {
-		case 0:
-			brushsize0Button->SetValue(true);
-			break;
-		case 1:
-			brushsize1Button->SetValue(true);
-			break;
-		case 2:
-			brushsize2Button->SetValue(true);
-			break;
-		case 4:
-			brushsize4Button->SetValue(true);
-			break;
-		case 6:
-			brushsize6Button->SetValue(true);
-			break;
-		case 8:
-			brushsize8Button->SetValue(true);
-			break;
-		case 11:
-			brushsize11Button->SetValue(true);
-			break;
-		default:
-			brushsize0Button->SetValue(true);
-			break;
-	}
-}
-
-void BrushSizePanel::OnClickCircleBrush(wxCommandEvent& event) {
-	g_palettes.ActivatePalette(GetParentPalette());
-	g_brush_manager.SetBrushShape(BRUSHSHAPE_CIRCLE);
-}
-
-void BrushSizePanel::OnClickSquareBrush(wxCommandEvent& event) {
-	g_palettes.ActivatePalette(GetParentPalette());
-	g_brush_manager.SetBrushShape(BRUSHSHAPE_SQUARE);
-}
-
-void BrushSizePanel::OnClickBrushSize(int which) {
-	g_palettes.ActivatePalette(GetParentPalette());
-	g_brush_manager.SetBrushSize(which);
-}
+#include "app/settings.h"
+#include "ui/gui_ids.h"
+#include "game/materials.h" // For g_materials and g_settings
 
 // ============================================================================
 // Tool Brush Panel
@@ -837,94 +477,4 @@ void BrushToolPanel::OnClickLockDoorCheckbox(wxCommandEvent& event) {
 
 	// save user preference
 	g_settings.setInteger(Config::DRAW_LOCKED_DOOR, event.IsChecked());
-}
-
-// ============================================================================
-// Brush Button
-
-BEGIN_EVENT_TABLE(BrushButton, ItemToggleButton)
-EVT_KEY_DOWN(BrushButton::OnKey)
-END_EVENT_TABLE()
-
-BrushButton::BrushButton(wxWindow* parent, Brush* _brush, RenderSize sz, uint32_t id) :
-	ItemToggleButton(parent, sz, uint16_t(0), id),
-	brush(_brush) {
-	ASSERT(sz != RENDER_SIZE_64x64);
-	ASSERT(brush);
-	SetSprite(brush->getLookID());
-	SetToolTip(wxstr(brush->getName()));
-}
-
-BrushButton::~BrushButton() {
-	////
-}
-
-void BrushButton::OnKey(wxKeyEvent& event) {
-	g_gui.AddPendingCanvasEvent(event);
-}
-
-// ============================================================================
-// Brush Thickness Panel
-
-BEGIN_EVENT_TABLE(BrushThicknessPanel, PalettePanel)
-#ifdef __WINDOWS__
-// This only works in wxmsw
-EVT_COMMAND_SCROLL_CHANGED(PALETTE_DOODAD_SLIDER, BrushThicknessPanel::OnScroll)
-#else
-EVT_COMMAND_SCROLL_TOP(PALETTE_DOODAD_SLIDER, BrushThicknessPanel::OnScroll)
-EVT_COMMAND_SCROLL_BOTTOM(PALETTE_DOODAD_SLIDER, BrushThicknessPanel::OnScroll)
-EVT_COMMAND_SCROLL_LINEUP(PALETTE_DOODAD_SLIDER, BrushThicknessPanel::OnScroll)
-EVT_COMMAND_SCROLL_LINEDOWN(PALETTE_DOODAD_SLIDER, BrushThicknessPanel::OnScroll)
-EVT_COMMAND_SCROLL_PAGEUP(PALETTE_DOODAD_SLIDER, BrushThicknessPanel::OnScroll)
-EVT_COMMAND_SCROLL_PAGEDOWN(PALETTE_DOODAD_SLIDER, BrushThicknessPanel::OnScroll)
-EVT_COMMAND_SCROLL_THUMBRELEASE(PALETTE_DOODAD_SLIDER, BrushThicknessPanel::OnScroll)
-#endif
-
-EVT_CHECKBOX(PALETTE_DOODAD_USE_THICKNESS, BrushThicknessPanel::OnClickCustomThickness)
-END_EVENT_TABLE()
-
-BrushThicknessPanel::BrushThicknessPanel(wxWindow* parent) :
-	PalettePanel(parent, wxID_ANY) {
-	wxSizer* thickness_sizer = newd wxBoxSizer(wxVERTICAL);
-
-	wxSizer* thickness_sub_sizer = newd wxBoxSizer(wxHORIZONTAL);
-	thickness_sub_sizer->Add(20, 10);
-	use_button = newd wxCheckBox(this, PALETTE_DOODAD_USE_THICKNESS, "Use custom thickness");
-	thickness_sub_sizer->Add(use_button);
-	thickness_sizer->Add(thickness_sub_sizer, 1, wxEXPAND);
-
-	slider = newd wxSlider(this, PALETTE_DOODAD_SLIDER, 5, 1, 10, wxDefaultPosition);
-	thickness_sizer->Add(slider, 1, wxEXPAND);
-
-	SetSizerAndFit(thickness_sizer);
-}
-
-BrushThicknessPanel::~BrushThicknessPanel() {
-	////
-}
-
-wxString BrushThicknessPanel::GetName() const {
-	return "Brush Thickness";
-}
-
-void BrushThicknessPanel::OnScroll(wxScrollEvent& event) {
-	static const int lookup_table[10] = { 1, 2, 3, 5, 8, 13, 23, 35, 50, 80 };
-	use_button->SetValue(true);
-
-	ASSERT(event.GetPosition() >= 1);
-	ASSERT(event.GetPosition() <= 10);
-
-	g_palettes.ActivatePalette(GetParentPalette());
-	g_brush_manager.SetBrushThickness(true, lookup_table[event.GetPosition() - 1], 100);
-}
-
-void BrushThicknessPanel::OnClickCustomThickness(wxCommandEvent& event) {
-	g_palettes.ActivatePalette(GetParentPalette());
-	g_brush_manager.SetBrushThickness(event.IsChecked());
-}
-
-void BrushThicknessPanel::OnSwitchIn() {
-	static const int lookup_table[10] = { 1, 2, 3, 5, 8, 13, 23, 35, 50, 80 };
-	g_palettes.ActivatePalette(GetParentPalette());
-	g_brush_manager.SetBrushThickness(lookup_table[slider->GetValue() - 1], 100);
 }
