+++ source/ui/games/game_panel.h
@@ -15,13 +15,48 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 //////////////////////////////////////////////////////////////////////
 
-#ifndef RME_PALETTE_COMMONS_H_
-#define RME_PALETTE_COMMONS_H_
+#ifndef RME_UI_GAMES_GAME_PANEL_H_
+#define RME_UI_GAMES_GAME_PANEL_H_
 
-#include "palette/palette_panel.h"
-#include "palette/brush_size_panel.h"
-#include "palette/brush_tool_panel.h"
-#include "palette/brush_thickness_panel.h"
-#include "palette/brush_button.h"
+#include "app/main.h"
+
+class GamePanel : public wxPanel {
+public:
+	GamePanel(wxWindow* parent, int width, int height);
+	virtual ~GamePanel();
+
+	void OnPaint(wxPaintEvent&);
+	void OnKeyDown(wxKeyEvent&);
+	void OnKeyUp(wxKeyEvent&);
+	void OnIdle(wxIdleEvent&);
+
+	void pause() {
+		paused_val = true;
+	}
+	void unpause() {
+		paused_val = false;
+	}
+	bool paused() const {
+		return paused_val || dead;
+	}
+
+protected:
+	virtual void Render(wxDC& pdc) = 0;
+	virtual void GameLoop(int time) = 0;
+	virtual void OnKey(wxKeyEvent& event, bool down) = 0;
+
+	virtual int getFPS() const = 0;
+
+protected:
+	wxStopWatch game_timer;
+
+private:
+	bool paused_val;
+
+	DECLARE_EVENT_TABLE()
+
+protected:
+	bool dead;
+};
 
 #endif
