+++ source/ui/map_window.cpp
@@ -222,3 +222,33 @@ void MapWindow::OnScrollPageUp(wxScrollEvent& event) {
 	}
 	Refresh();
 }
+
+BEGIN_EVENT_TABLE(MapWindow, wxPanel)
+EVT_SIZE(MapWindow::OnSize)
+
+EVT_COMMAND_SCROLL_TOP(MAP_WINDOW_HSCROLL, MapWindow::OnScroll)
+EVT_COMMAND_SCROLL_BOTTOM(MAP_WINDOW_HSCROLL, MapWindow::OnScroll)
+EVT_COMMAND_SCROLL_THUMBTRACK(MAP_WINDOW_HSCROLL, MapWindow::OnScroll)
+EVT_COMMAND_SCROLL_LINEUP(MAP_WINDOW_HSCROLL, MapWindow::OnScrollLineUp)
+EVT_COMMAND_SCROLL_LINEDOWN(MAP_WINDOW_HSCROLL, MapWindow::OnScrollLineDown)
+EVT_COMMAND_SCROLL_PAGEUP(MAP_WINDOW_HSCROLL, MapWindow::OnScrollPageUp)
+EVT_COMMAND_SCROLL_PAGEDOWN(MAP_WINDOW_HSCROLL, MapWindow::OnScrollPageDown)
+
+EVT_COMMAND_SCROLL_TOP(MAP_WINDOW_VSCROLL, MapWindow::OnScroll)
+EVT_COMMAND_SCROLL_BOTTOM(MAP_WINDOW_VSCROLL, MapWindow::OnScroll)
+EVT_COMMAND_SCROLL_THUMBTRACK(MAP_WINDOW_VSCROLL, MapWindow::OnScroll)
+EVT_COMMAND_SCROLL_LINEUP(MAP_WINDOW_VSCROLL, MapWindow::OnScrollLineUp)
+EVT_COMMAND_SCROLL_LINEDOWN(MAP_WINDOW_VSCROLL, MapWindow::OnScrollLineDown)
+EVT_COMMAND_SCROLL_PAGEUP(MAP_WINDOW_VSCROLL, MapWindow::OnScrollPageUp)
+EVT_COMMAND_SCROLL_PAGEDOWN(MAP_WINDOW_VSCROLL, MapWindow::OnScrollPageDown)
+
+EVT_BUTTON(MAP_WINDOW_GEM, MapWindow::OnGem)
+END_EVENT_TABLE()
+
+BEGIN_EVENT_TABLE(MapScrollBar, wxScrollBar)
+EVT_KEY_DOWN(MapScrollBar::OnKey)
+EVT_KEY_UP(MapScrollBar::OnKey)
+EVT_CHAR(MapScrollBar::OnKey)
+EVT_SET_FOCUS(MapScrollBar::OnFocus)
+EVT_MOUSEWHEEL(MapScrollBar::OnWheel)
+END_EVENT_TABLE()
